requirements.txt
openai>=1.0.0
elevenlabs>=0.2.0
mcp>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
websockets>=11.0.0
pyaudio>=0.2.11
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pydub>=0.25.1
requests>=2.31.0
pygame>=2.0.0
numpy>=1.21.0
